+++ src/Tests/Analyzers.Tests/RCS1235OptimizeMethodCallTests.cs
@@ -406,6 +406,91 @@ class C
         }
 
         [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.OptimizeMethodCall)]
+        public async Task Test_OptimizeAdd()
+        {
+            await VerifyDiagnosticAndFixAsync(@"
+using System.Collections.Generic;
+
+class C
+{
+    void M()
+    {
+        var items = new List<string>();
+
+        foreach (var item in items)
+        {
+            items.[|Add|](item);
+        }
+    }
+}
+", @"
+using System.Collections.Generic;
+
+class C
+{
+    void M()
+    {
+        var items = new List<string>();
+
+        items.AddRange(items);
+    }
+}
+");
+        }
+
+        [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.OptimizeMethodCall)]
+        public async Task Test_OptimizeAdd_EmbeddedStatement()
+        {
+            await VerifyDiagnosticAndFixAsync(@"
+using System.Collections.Generic;
+
+class C
+{
+    void M()
+    {
+        var items = new List<string>();
+
+        foreach (var item in items)
+            items.[|Add|](item);
+    }
+}
+", @"
+using System.Collections.Generic;
+
+class C
+{
+    void M()
+    {
+        var items = new List<string>();
+
+        items.AddRange(items);
+    }
+}
+");
+        }
+
+        [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.OptimizeMethodCall)]
+        public async Task TestNoDiagnostic_OptimizeAdd_Await()
+        {
+            await VerifyNoDiagnosticAsync(@"
+using System.Threading.Tasks;
+using System.Collections.Generic;
+
+class C
+{
+    async Task M()
+    {
+        var items = new List<string>();
+        IAsyncEnumerable<string> items2 = null;
+
+        await foreach (var item in items2)
+            items.Add(item);
+    }
+}
+");
+        }
+
+        [Fact, Trait(Traits.Analyzer, DiagnosticIdentifiers.OptimizeMethodCall)]
         public async Task TestNoDiagnostic_CallCompareOrdinalInsteadOfCompare_NotStringComparisonOrdinal()
         {
             await VerifyNoDiagnosticAsync(@"